--- TTWarsCore/FindElements/Building.cs
@@ -1,18 +1,0 @@
-﻿using HtmlAgilityPack;
-using System;
-
-namespace TTWarsCore.FindElements
-{
-    public static class Building
-    {
-        public static HtmlNode GetResourceField(HtmlDocument doc, int index)
-        {
-            throw new NotImplementedException();
-        }
-
-        public static HtmlNode GetBuilding(HtmlDocument doc, int index)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
