+++ TTWarsCore/Parsers/BuildingTabParser.cs
@@ -1,19 +1,20 @@
 ﻿using HtmlAgilityPack;
+using ModuleCore.Parser;
 using System.Collections.Generic;
 using System.Linq;
 
 namespace TTWarsCore.Parsers
 {
-    public static class BuildingTab
+    public class BuildingTabParser : IBuildingTabParser
     {
-        public static List<HtmlNode> GetBuildingTabNodes(HtmlDocument doc)
+        public List<HtmlNode> GetBuildingTabNodes(HtmlDocument doc)
         {
             var navigationBar = doc.DocumentNode.Descendants("div").FirstOrDefault(x => x.HasClass("contentNavi") && x.HasClass("subNavi"));
-            var navNodes = navigationBar.Descendants("div").Where(x => x.HasClass("container"));
+            var navNodes = navigationBar.Descendants("a").Where(x => x.HasClass("tabItem"));
             return navNodes.ToList();
         }
 
-        public static bool IsCurrentTab(HtmlNode tabNode)
+        public bool IsCurrentTab(HtmlNode tabNode)
         {
             return tabNode.HasClass("active");
         }
