+++ TTWarsCore/Parsers/FarmListParser.cs
@@ -1,35 +1,36 @@
 ﻿using HtmlAgilityPack;
+using ModuleCore.Parser;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace TravianOfficialCore.Parsers
+namespace TTWarsCore.Parsers
 {
-    public static class FarmList
+    public class FarmListParser : IFarmListParser
     {
-        public static List<HtmlNode> GetFarmNodes(HtmlDocument doc)
+        public List<HtmlNode> GetFarmNodes(HtmlDocument doc)
         {
             var raidList = doc.GetElementbyId("raidList");
             if (raidList is null) return new();
-            var fls = raidList.Descendants("div").Where(x => x.HasClass("raidList"));
+            var fls = raidList.ChildNodes.Where(x => x.Id.StartsWith("list"));
 
             return fls.ToList();
         }
 
-        public static string GetName(HtmlNode node)
+        public string GetName(HtmlNode node)
         {
-            var flName = node.Descendants("div").FirstOrDefault(x => x.HasClass("listName"));
+            var flName = node.Descendants("div").FirstOrDefault(x => x.HasClass("listTitleText"));
             if (flName is null) return null;
             return flName.InnerText.Trim();
         }
 
-        public static int GetId(HtmlNode node)
+        public int GetId(HtmlNode node)
         {
-            var id = node.GetAttributeValue("data-listid", "0");
+            var id = node.Id;
             var value = new string(id.Where(c => char.IsDigit(c)).ToArray());
             return int.Parse(value);
         }
 
-        public static int GetNumOfFarms(HtmlNode node)
+        public int GetNumOfFarms(HtmlNode node)
         {
             var slotCount = node.Descendants("span").FirstOrDefault(x => x.HasClass("raidListSlotCount"));
             if (slotCount is null) return 0;
