+++ TTWarsCore/TTWarsCore.csproj
@@ -10,4 +10,8 @@
     <PackageReference Include="Selenium.WebDriver" Version="4.7.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\ModuleCore\ModuleCore.csproj" />
+  </ItemGroup>
+
 </Project>
