+++ TravianOfficialCore/Parsers/VillageFieldParser.cs
@@ -1,31 +1,38 @@
 ﻿using HtmlAgilityPack;
+using ModuleCore.Parser;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace TTWarsCore.Parsers
+namespace TravianOfficialCore.Parsers
 {
-    public static class VillageFields
+    public class VillageFieldParser : IVillageFieldParser
     {
-        public static List<HtmlNode> GetResourceNodes(HtmlDocument doc)
+        public List<HtmlNode> GetNodes(HtmlDocument doc)
         {
-            var villageMapNode = doc.GetElementbyId("village_map");
-            if (villageMapNode is null) return new();
+            var resourceFieldContainerNode = doc.GetElementbyId("resourceFieldContainer");
+            if (resourceFieldContainerNode is null) return new();
 
-            return villageMapNode.Descendants("div").Where(x => !x.HasClass("labelLayer")).ToList();
+            return resourceFieldContainerNode.ChildNodes.Where(x => x.HasClass("level")).ToList();
         }
 
-        public static int GetId(HtmlNode node)
+        public HtmlNode GetNode(HtmlDocument doc, int index)
+        {
+            return doc.DocumentNode.Descendants("a").FirstOrDefault(x => x.HasClass($"buildingSlot{index}"));
+        }
+
+        public int GetId(HtmlNode node)
         {
             var classess = node.GetClasses();
-            var needClass = classess.FirstOrDefault(x => x.StartsWith("aid"));
+            var needClass = classess.FirstOrDefault(x => x.StartsWith("buildingSlot"));
             if (string.IsNullOrEmpty(needClass)) return -1;
             var strResult = new string(needClass.Where(c => char.IsDigit(c)).ToArray());
             if (string.IsNullOrEmpty(strResult)) return -1;
+
             return int.Parse(strResult);
         }
 
-        public static int GetType(HtmlNode node)
+        public int GetBuildingType(HtmlNode node)
         {
             var classess = node.GetClasses();
             var needClass = classess.FirstOrDefault(x => x.StartsWith("gid"));
@@ -36,17 +43,18 @@ namespace TTWarsCore.Parsers
             return int.Parse(strResult);
         }
 
-        public static int GetLevel(HtmlNode node)
+        public int GetLevel(HtmlNode node)
         {
             var classess = node.GetClasses();
             var needClass = classess.FirstOrDefault(x => x.StartsWith("level") && !x.Equals("level"));
             if (string.IsNullOrEmpty(needClass)) return -1;
             var strResult = new string(needClass.Where(c => char.IsDigit(c)).ToArray());
+            if (string.IsNullOrEmpty(strResult)) return -1;
 
             return int.Parse(strResult);
         }
 
-        public static bool IsUnderConstruction(HtmlNode node)
+        public bool IsUnderConstruction(HtmlNode node)
         {
             return node.GetClasses().Contains("underConstruction");
         }
